ttf files descomposition and rasterization

+++ src/primitives/file_processing/readers/sequential_file_readed/read16.cpp
@@ -1,8 +1,7 @@
 #include "../../../../../headers/primitives/file_processing.hpp"
 
 uint16_t SequentialFileReader::read16 () {
-  uint32_t res = this->read8() | 
-    (this->read8() << 8);
+  uint16_t res = this->read8() | (this->read8() << 8);
   if constexpr (std::endian::native == std::endian::little)
     return std::byteswap(res);
   else
